+++ NEWS.md
@@ -1,3 +1,7 @@
+## RFCCA 1.0.4
+* Added 'samptype' and 'sampsize' arguments to rfcca() function.
+* Fixed the displaying with 'ndisp' argument in plot.vimp.rfcca() function.
+
 ## RFCCA 1.0.3
 * Fixed the BOP construction for test observations.
 * Updated README file.